// ui/src/api/generated/model/getInspectionKpisManagerApiInspectionsKpisManagerGetPeriod.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * Whirlpool Inspection Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.0.3
 */

export type GetInspectionKpisManagerApiInspectionsKpisManagerGetPeriod = typeof GetInspectionKpisManagerApiInspectionsKpisManagerGetPeriod[keyof typeof GetInspectionKpisManagerApiInspectionsKpisManagerGetPeriod];


export const GetInspectionKpisManagerApiInspectionsKpisManagerGetPeriod = {
  custom: 'custom',
  today: 'today',
  yesterday: 'yesterday',
  week: 'week',
  month: 'month',
} as const;